refac: created infinite list file
